requirements.txt
bleak>=0.21.1
flipperzero-protobuf-py>=0.17.2
asyncio>=3.4.3
typing>=3.7.4.3
pyyaml>=6.0.1
ttkthemes>=3.2.2